02_mcp_adk_client/requirements.txt
mcp>=1.2.0
google-adk>=0.1.0
rich
anyio
python-dotenv
